refactoring/#90 [Backend] Adicionado arquivo DepencyInjection em pasta IoC
